+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## [1.0.0] - 
+## [1.1.0] - 25/87/2023
+### Added
+- ClearPlayers() to Manager
+- spawnOnStart to player spawner
+
+## [1.0.0] - 03/07/2023
 ### Added
 - First public release